BACK-PARKVISION-P2M/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
passlib[bcrypt]
python-dotenv
psycopg2-binary
email-validator
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
pydantic
python-multipart